+++ src/merge/mkvmerge.cpp
@@ -416,7 +416,7 @@ print_capabilities() {
 
 static std::string
 guess_mime_type_and_report(std::string file_name) {
-  auto mime_type = mtx::mime::guess_type(file_name, true);
+  auto mime_type = mtx::mime::guess_type_for_file(file_name);
   if (mime_type != "") {
     mxinfo(fmt::format(Y("Automatic MIME type recognition for '{0}': {1}\n"), file_name, mime_type));
     return mime_type;