+++ src/mkvtoolnix-gui/header_editor/tab.cpp
@@ -792,7 +792,7 @@ Tab::createAttachmentFromFile(QString const &fileName) {
   if (!content)
     return {};
 
-  auto mimeType   = mtx::mime::guess_type(to_utf8(fileName), true);
+  auto mimeType   = mtx::mime::guess_type_for_file(to_utf8(fileName));
   auto uid        = create_unique_number(UNIQUE_ATTACHMENT_IDS);
   auto fileData   = new KaxFileData;
   auto attachment = KaxAttachedPtr{